+++ app/workers/recurring_pod_check.rb
@@ -1,11 +1,6 @@
-
 module Workers
   class RecurringPodCheck < Base
-    include Sidetiq::Schedulable
-
     sidekiq_options queue: :maintenance
-
-    recurrence { daily }
 
     def perform
       Pod.check_all!