+++ config/initializers/sidekiq.rb
@@ -52,3 +52,9 @@ end
 Sidekiq.configure_client do |config|
   config.redis = AppConfig.get_redis_options
 end
+
+schedule_file = "config/schedule.yml"
+
+if File.exist?(schedule_file) && Sidekiq.server?
+  Sidekiq::Cron::Job.load_from_hash YAML.load_file(schedule_file)
+end